+++ p_telept.c
@@ -31,9 +31,7 @@
 //
 // TELEPORTATION
 //
-int EV_Teleport(line_t *line,
-				int side,
-				mobj_t *thing)
+int EV_Teleport(line_t *line, byte side, mobj_t *thing)
 {
 	int i;
 	int tag;
